// front_silant/src/providers/Api/models/ListMachine.tsx
export interface ListMachine{
            filter_data:{
                controlled_bridge_models: object
                driving_bridge_models: object
                engine_models:object
                machine_models: object
                transmission_models: object
        },
            machine_list_data:object,
            users_data: object,
            services_data: object


}